pulsevideosrc: Revamp cancellation
`socketsrc` calls `on_socket_eos` to get a new socket.  This is a blocking
operation, but `socketsrc` did not treat it like that.  With its other
blocking operations `socketsrc` would pass in a `GCancellable` to allow it
to be cancelled as necessary, but it would not do that for `on_socket_eos`.
Cancellation used to rely on the `GCancellable` in `pulsevideosrc` being
in-sync with the one in `socketsrc`.

With this commit we ditch the `GCancellable` from pulsevideosrc.
pulsevideosrc will borrow it from socketsrc during `on_socket_eos` instead.
This is simpler because there is less coordination.

Note that `gst_pulsevideo_src_start` now passes `NULL` for cancellable
into `gst_pulsevideo_src_reinit`.  This is not a behavioural change - in
fact the initial connect was never cancellable, despite what you might
expect because `g_cancellable_cancel` can never get called.

This change also highlights that the initial connect is not cancellable,
despite `GCancellable` being used because there is no point at which
cancel can be called.  It was called in the transition from PAUSED to
READY, but that can't happen if we're currently transitioning from
READY to PAUSED and we're blocked on connect.  We'll solve this in a future
commit.

Regarding keeping the two cancellables in-sync: I'm not sure, but this may
have caused teardown hangs or other weirdness in the case that the
pulsevideo server is in a bad state while we shut down.  Previously the
sequence could be:

1. Pulsevideo server is having difficulty
2. `socketsrc` calls `on_socket_eos` - which blocks.  This is called in the
   stream thread.
3. Pipeline is tearing down and state change PAUSED -> READY is called on
   pulsevideosrc from the main thread.
4. `pulsevideosrc` calls state change on `socketsrc` which blocks on the
   stream thread.  This means that the main thread is now blocked on the
   stream thread which is blocked on `on_socket_eos` which is blocked on
   upstream pulsevideo server.
5. `pulsevideosrc` cancels its `GCancellable` unblocking everything - but
   we don't get this far because we're blocked at (4).

One fix would have been to have pulsevideosrc cancel its cancellable before
calling into socketsrc, but I think this is cleaner.

This should allow for more orderly shutdowns.