pulsevideosrc: Make startup cancellable
As described in the previous commit, currently it's not possible that we
can cancel startup if we're blocked in `gst_pulsevideo_src_reinit`.  We may
get blocked for an arbitrarily long time if the pulsevideo server is in a
bad state.

With this change we don't block on connecting upstream in the state change
function, we instead go to PAUSED state directly and rely on the connection
happening in response to a `on_socket_eos` signal from the stream thread.
This means it will be interruptable because the socketsrc fill function is.

This may affect caps negotiation - the final caps will only be available
once we're actually in the PAUSED state, rather than being set during the
state change.